requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.10.1
pydantic-settings==2.6.1
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic[email]